requirements.txt
streamlit>=1.35.0
mlx-whisper>=0.3.0
pyannote.audio>=3.1.0
pyaudio>=0.2.14
ollama>=0.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
